Extract `CallWithChatComposite` browser test app to external folder (#2263)
