+++ change/@internal-react-composites-b0de9eb0-0366-442e-8447-14b090ac1de1.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "Internal reorganization of browser test folders",
+  "packageName": "@internal/react-composites",
+  "email": "82062616+prprabhu-ms@users.noreply.github.com",
+  "dependentChangeType": "none"
+}
