+++ packages/react-composites/tests/app/callwithchat/index.tsx
@@ -11,9 +11,9 @@ import {
   CallWithChatAdapterState,
   CallWithChatComposite,
   useAzureCommunicationCallWithChatAdapter
-} from '../../../../src';
-import { IDS } from '../../common/constants';
-import { initializeIconsForUITests, isMobile, verifyParamExists } from '../../common/testAppUtils';
+} from '../../../src';
+import { IDS } from '../../browser/common/constants';
+import { initializeIconsForUITests, isMobile, verifyParamExists } from '../../browser/common/testAppUtils';
 import memoizeOne from 'memoize-one';
 import { fromFlatCommunicationIdentifier } from '@internal/acs-ui-common';
