+++ packages/react-composites/tests/app/callwithchat/webpack.config.js
@@ -4,8 +4,8 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 
 module.exports = () => {
-  const babelConfig = require('../../../../.babelrc.js');
-  const commonConfig = require('../../common/testapp.webpack.config')(__dirname, babelConfig);
+  const babelConfig = require('../../../.babelrc.js');
+  const commonConfig = require('..//testapp.webpack.config')(__dirname, babelConfig);
 
   return commonConfig;
 };
