--- packages/react-composites/tests/browser/common/fonts/readme.md
@@ -1,4 +1,0 @@
-# E2E Test fonts
-
-These are the required fonts used by fluent that fluent would regularly download from their server (<https://static2.sharepointonline.com/files/fabric/assets/fonts/>).
-The reason to have these included locally is to avoid unnecessary outgoing network calls in e2e tests.
