--- packages/react-composites/tests/browser/common/index.html
@@ -1,68 +1,0 @@
-<!-- Copyright (c) Microsoft Corporation. -->
-<!-- Licensed under the MIT license. -->
-
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="description" content="Composite - Test App" />
-    <title>Composite - Test App</title>
-
-    <!-- Ensure that fonts have loaded before tests are run -->
-    <link rel="preload" href="./fonts/segoeui-regular.woff2" as="font" crossorigin="anonymous" />
-    <link rel="preload" href="./fonts/segoeui-semibold.woff2" as="font" crossorigin="anonymous" />
-    <link rel="preload" href="./fonts/segoeui-bold.woff2" as="font" crossorigin="anonymous" />
-    <style>
-      @font-face {
-        font-weight: 400;
-        src: url('./fonts/segoeui-regular.woff2') format('woff2');
-        font-family: 'Segoe UI';
-      }
-      @font-face {
-        font-weight: 600;
-        src: url('./fonts/segoeui-semibold.woff2') format('woff2');
-        font-family: 'Segoe UI';
-      }
-      @font-face {
-        font-weight: 700;
-        src: url('./fonts/segoeui-bold.woff2') format('woff2');
-        font-family: 'Segoe UI';
-      }
-    </style>
-
-    <style>
-      /* Important: For ensuring that blinking cursor doesn't get captured in
-         snapshots and cause a diff in subsequent tests. */
-      * {
-        caret-color: transparent !important;
-      }
-
-      /* Help with inconsistent font rendering */
-      * {
-        text-rendering: geometricprecision !important;
-      }
-
-      /* Prevent animations. waitForElementState('stable') is not working for opacity
-         animation https://github.com/microsoft/playwright/issues/4055#issuecomment-777697079.
-         Examples of components that currently animate are button flyouts and button tooltips */
-      *,
-      *::before,
-      *::after {
-        transition: none !important;
-        -webkit-transition: none !important;
-        -moz-transition: none !important;
-        -ms-transition: none !important;
-        -o-transition: color 0 ease-in !important;
-        animation: none !important;
-      }
-    </style>
-  </head>
-
-  <body>
-    <main id="root" class="Root">
-      <!-- Anything in here is replaced by react when react mounts and completes the first render pass -->
-      Loading...
-    </main>
-  </body>
-</html>
